add authorized blog deleting

+++ backend/controllers/blogs.js
@@ -1,24 +1,14 @@
 const blogsRouter = require('express').Router()
-// const jwt = require('jsonwebtoken')
-const helper = require('../tests/test_helper')
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
-
-
-// const getTokenFrom = request => {
-//   const authorization = request.get('authorization')
-//   if (authorization && authorization.startsWith('Bearer ')) {
-//     return authorization.replace('Bearer ', '')
-//   }
-//   return null
-// }
 
 
 // fetching all blogs
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
     .find({})
-    .populate('user', { username: 1, id: 1})
+    .populate('user', { username: 1, id: 1 })
   response.json(blogs)
 })
 
@@ -26,21 +16,19 @@ blogsRouter.get('/', async (request, response) => {
 // adding new blog
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  const user = await User.findById(body.user)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
-  // const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-  // if (!decodedToken.id) {
-  //   return response.status(401).json({ error: 'token invalid' })
-  // }
-  // const user = await User.findById(decodedToken.id)
-
-  // if (!user) {
-  //   return response.status(400).json({ error: 'userId missing or not valid' })
-  // }
+  if (!user) {
+    return response.status(400).json({ error: 'userId missing or not valid' })
+  }
 
   const blog = new Blog({
     title: body.title,
-    author: body.author,
+    author: user.name,
     url: body.url,
     likes: body.likes || 0,
     user: user.id
@@ -55,8 +43,24 @@ blogsRouter.post('/', async (request, response) => {
 
 // deleting a blog
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  if (!user) {
+    return response.status(400).json({ error: 'userId missing or not valid' })
+  }
+
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(400).json({ error: 'User is not the creator of the blog' })
+  }
 })
